Add Kafka
Kafka is used by the TV Kitchen, which means this project is going to
rely on it as well. This sets up a docker compose file for Kafka which
should make deployment much easier.

We don't want to use a hosted version of Kafka because of the bandwidth
considerations.

Issue #1